🐛 check files existence before removal

+++ src/remove.rs
@@ -38,7 +38,9 @@ fn remove_migration_file(last_migration: &SurqlFile) -> Result<()> {
 
     let file_path = migrations_path.join(&last_migration.full_name);
 
-    std::fs::remove_file(file_path)?;
+    if file_path.exists() {
+        std::fs::remove_file(file_path)?;
+    }
 
     Ok(())
 }
